+++ packages/Peptides/src/widgets/selection.ts
@@ -50,5 +50,13 @@ export function getSelectionWidget(table: DG.DataFrame, model: PeptidesModel): H
 
   const gridHost = ui.box(grid.root);
   gridHost.style.marginLeft = '0px';
+  setTimeout(() => {
+    for (let gridColIdx = 1; gridColIdx < sourceGrid.columns.length; gridColIdx++) {
+      const gridCol = sourceGrid.columns.byIndex(gridColIdx)!;
+      if (!gridCol.visible)
+        continue;
+      grid.col(gridCol.name)!.width = gridCol.width;
+    }
+  }, 500);
   return gridHost;
 }